feat:냉파 점수 산정 내역 조회 API 구현(#94)

+++ src/main/java/com/naengpa/naengpamasterbackend/score/entity/Score.java
@@ -23,7 +23,7 @@ public class Score {
     private Long memberId;
 
     @Column(name = "grade_id", nullable = false)
-    private int gradeId;
+    private Long gradeId;
 
     @Column(name = "score")
     private int score;
